SimpleView polyfill refactor (#123)
This PR does a medium size refactoring on simpleView polyfill that I consider need to expand to the rest of polyfills:

SimpleView.test.ts
Currently the tests verify that a function can be override by the polyfills. The issue is that the actual content of the file constantly changes. This PR forks the behavior as follows:
If run in the server: It will continue working as today, verifying that a static text can be patched.
If run locally: it will verify if the actual files exist in disk, if they do it will run the tests again the file contents, and it will fail if the contents do not match. This will make patching and spotting differences easier by reducing turnaround time
PostBuild.ts:
Currently "Release" and "Debug" mode are treated as one, this creates issues as the files on DevTools side tend to be moved or refactored, this makes the patching process a manual one. This PR splits the logic between Release and Debug, so if a release file that is needed to be patched is not present it will now throw an error.

launch.json
As part of the change in postbuild.ts, the user can now indicate if they are running a "Release" version (following Contributing.md instructions or if they are doing a development version (full chromium repo)

The rest of the changes are to support the scenarios mentioned above.