Removing debug mode. (#183)
Currently there is a configuration named "debug mode" which targets the source files of the Edge Chromium repo (not the output files). This was done to enable coding with the Devtools source files. Unfortunately this created a dependency to the source files itself (requiring the full 2gb file). This PR removes this dependency, any changes required on Devtools should be done in Devtools repo or patched through a "polyfill"

Addresses #169