[Screencast]: Fix blurriness on high-dpi displays (#574)
Previously the screencast assumed a 100% zoom factor, leading it to be scaled-up on higher DPI displays and become blurry.

Fixed by scaling by the `devicePixelRatio` when requesting the screencast, but keeping the viewport, image size, and inputs all mapped back to layout pixels. The scaled result is floored to ensure the requested dimensions are whole pixels (as required by CDP).