Updating AI Key (#2743)
Upgrading typescript and vscode-extension-telemetry (#2704)

Upgraded extension telemetry handler. The new version is not complatible with previous AIF keys but 1DS. From previous documentation it seems the new one has to be updated, otherwise the application will not be registered and it will not work.

https://github.com/microsoft/vscode-azure-ext-sdk/blob/main/readme.md
points to
https://github.com/Microsoft/vscode-docker/blob/main/package.json

Which contains the new key.